requirements.txt
anyio              ==4.6.2.post1
certifi            == 2024.8.30
h11                == 0.14.0
httpcore           == 1.0.6
httpx              == 0.27.2
idna               == 3.10
pip                == 24.3.1
pyaes              == 1.6.1
pyasn1             == 0.6.1
python-dotenv      == 1.0.1
python-telegram-bot== 21.7
rsa                == 4.9
sniffio            == 1.3.1
Telethon           == 1.37.0
tinydb             == 4.8.2